% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnorm.gibbs.R
\name{bnorm.gibbs}
\alias{bnorm.gibbs}
\title{Gibbs Sampler: Bivariate Normal}
\usage{
bnorm.gibbs(x0 = c(-5, 5), rho = 0.8, n = 500)
}
\arguments{
\item{x0}{initial states}

\item{rho}{correlation coefficient}

\item{n}{number of generated points MC}
}
\description{
Gibbs Sampler: Bivariate Normal
}
\examples{
bnorm.gibbs(x0 = c(-5, 5), rho = 0.8, n = 500)
}
