% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnorm.mh.R
\name{bnorm.mh}
\alias{bnorm.mh}
\title{Metropolis Hasting: Bivariate Normal}
\usage{
bnorm.mh(x0 = c(-5, 5), rho = 0.8, n = 500, FUN)
}
\arguments{
\item{x0}{initial states}

\item{rho}{correlation coefficient}

\item{n}{number of generated points MC}

\item{FUN}{dbnorm}
}
\description{
Metropolis Hasting: Bivariate Normal
}
\examples{
dbnorm = function(x, d) exp(-t(x) \%*\% d \%*\% x/2)
bnorm.mh(x0=c(-5,5), rho = 0.8, n = 500, FUN = dbnorm)
}
