% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.est.R
\name{bt.est}
\alias{bt.est}
\title{Bootstrap Estimates}
\usage{
bt.est(d, R = 1000, T)
}
\arguments{
\item{d}{data}

\item{R}{number of replications}

\item{T}{statistic:thetahat (mean, variance, ...)}
}
\description{
Bootstrap Estimates
}
\examples{
air=c(3,5,7,18,43,85,91,98,100,130,230,487)
logmean=function(x) log(mean(x))
ts=sort(bt.est(d=air, R=1000, T=logmean))
}
