% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bv.boot.R
\name{bv.boot}
\alias{bv.boot}
\title{Parametric Bootstrap}
\usage{
bv.boot(x, R = 1000, T)
}
\arguments{
\item{x}{data}

\item{R}{number of replications}

\item{T}{statistic:thetahat (mean, variance, ...)}
}
\description{
Parametric Bootstrap
}
\examples{
logmean=function(x) log(mean(x))
boeing=c(3,5,7,18,43,85,91,98,100,130,230,487)
bv.boot(boeing, R=1000, T=logmean)[1:2]
}
