% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bv.npboot.R
\name{bv.npboot}
\alias{bv.npboot}
\title{NonParametric Bootstrap}
\usage{
bv.npboot(x, R = 1000, T)
}
\arguments{
\item{x}{data}

\item{R}{number of replications}

\item{T}{statistic:thetahat (mean, variance, ...)}
}
\description{
NonParametric Bootstrap
}
\examples{
 logmean=function(x) log(mean(x))
air=c(3,5,7,18,43,85,91,98,100,130,230,487)
bv.npboot(air, R=1000, T=logmean)[1:2]
}
