% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Variance Reduction Method}
\usage{
cv(n = 10, a = 0, b = 1)
}
\arguments{
\item{n}{number of generated points}

\item{a}{lower limit}

\item{b}{upper limit}
}
\description{
Variance Reduction Method
}
\examples{
cv(n=1000, a=0, b=1)*4
}
