% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.R
\name{gsm}
\alias{gsm}
\title{Monte Carlo Integration}
\usage{
gsm(h, f, rf, n = 1)
}
\arguments{
\item{h}{function}

\item{f}{function}

\item{rf}{function}

\item{n}{number of generated points}
}
\description{
Monte Carlo Integration
}
\examples{
h=function(x) sqrt(1-x^2)
f=function(x) 2*x
rf=function(n) sqrt(runif(n))
gsm(h, f, rf, n = 1e5)*4
}
