% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{hm}
\alias{hm}
\title{Monte Carlo Integration}
\usage{
hm(h, a = 0, b = 1, c = 1, n = 1)
}
\arguments{
\item{h}{function}

\item{a}{lower limit}

\item{b}{upper limit}

\item{c}{max value of h function}

\item{n}{number of generated points}
}
\description{
Monte Carlo Integration
}
\examples{
h=function(x) sqrt(1-x^2)
hm(h,a=0,b=1,c=1,n=1e5)*4
}
