% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS.R
\name{IS}
\alias{IS}
\title{Variance Reduction Method}
\usage{
IS(n = 1, alpha = 1, beta = 1.3)
}
\arguments{
\item{n}{number of generated points}

\item{alpha}{shape parameter beta instrumental density}

\item{beta}{shape parameter beta instrumental density}
}
\description{
Variance Reduction Method
}
\examples{
IS(n=10000, alpha=1, beta=1.3)*4
}
