% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mchain.R
\name{mchain}
\alias{mchain}
\title{Simulating a Markov Chain}
\usage{
mchain(n = 10, x0 = 2, P)
}
\arguments{
\item{n}{number of generated points MC}

\item{x0}{initial states}

\item{P}{transition matrix}
}
\description{
Simulating a Markov Chain
}
\examples{
P1=rbind(c(3,1)/4, c(3,2)/5)
mchain(10, 2, P1)
}
