% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mchain.stdnorm.R
\name{mchain.stdnorm}
\alias{mchain.stdnorm}
\title{Chain for Standard Normal Distribution}
\usage{
mchain.stdnorm(x0 = -5, n = 200)
}
\arguments{
\item{x0}{initial states}

\item{n}{number of generated points MC}
}
\description{
Chain for Standard Normal Distribution
}
\examples{
mc=mchain.stdnorm(x0=-5, n=200)
}
