% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs.R
\name{pbs}
\alias{pbs}
\title{Parametric Bootstrap Test}
\usage{
pbs(bdata, R = 1000)
}
\arguments{
\item{bdata}{data}

\item{R}{number of replications}
}
\description{
Parametric Bootstrap Test
}
\details{
function rmvnorm requires to install package mvtnorm
}
