% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.R
\name{perm}
\alias{perm}
\title{Permutation Test}
\usage{
perm(bdata, R = 1000)
}
\arguments{
\item{bdata}{data}

\item{R}{number of replications}
}
\description{
Permutation Test
}
\examples{
perm(bdata=cars, R=1000)
}
