% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.n.R
\name{pi.n}
\alias{pi.n}
\title{Distribution of Xn}
\usage{
pi.n(n = 3, P, pi0)
}
\arguments{
\item{n}{number of generated points MC}

\item{P}{transition matrix}

\item{pi0}{initial distribution}
}
\description{
Distribution of Xn
}
\examples{
n=3
P= rbind(c(0.6, 0.4, 0), c(0.2, 0.6, 0.2), c(0, 0.4, 0.6))
pi0=rep(1/3, 3)
pi.n(n, P, pi0)
}
