% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa}
\alias{sa}
\title{Simulated Annealing Algorithm}
\usage{
sa(n = 5000, x = c(-1.9, 1.9))
}
\arguments{
\item{n}{number of generated points MC}

\item{x}{grid of measurements}
}
\description{
Simulated Annealing Algorithm
}
\details{
h=function(x){ if(is.vector(x)) dim(x)=c(1,2) cos(10*x[,1]^2)-5*x[,1]^2+cos(10*x[,2]^2)-5*x[,2]*2}

uwalk = function(x, step = c(0.1, 0.1), valid) { xn = x + runif(length(x), -1, 1) * step while (!valid(xn)) xn = x + runif(length(x),-1, 1) * step xn}

hv = function(x) !any(x <= -2 | x >= 2)

tempf = function(n) 0.995^(1:n) * 10000
}
