% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm}
\alias{sm}
\title{Monte Carlo Integration}
\usage{
sm(h, a = 0, b = 1, n = 1)
}
\arguments{
\item{h}{function}

\item{a}{lower limit}

\item{b}{upper limit}

\item{n}{number of generated points}
}
\description{
Monte Carlo Integration
}
\examples{
h=function(x) sqrt(1-x^2)
sm(h,a=0,b=1,n=1e5)*4
}
